+++ src/compas/geometry/hull/hull_numpy.py
@@ -35,47 +35,7 @@ def convex_hull_numpy(points):
 
     Examples
     --------
-    .. code-block:: python
-
-        import random
-
-        from compas.datastructures import Mesh
-
-        from compas.geometry import distance_point_point
-        from compas.geometry import convex_hull_numpy
-        from compas.topology import unify_cycles
-
-        from compas_viewers import MeshViewer
-
-        radius = 5
-        origin = (0., 0., 0.)
-        count = 0
-        points = []
-
-        while count < 10:
-            x = (random.random() - 0.5) * radius * 2
-            y = (random.random() - 0.5) * radius * 2
-            z = (random.random() - 0.5) * radius * 2
-            pt = x, y, z
-
-            if distance_point_point(origin, pt) <= radius:
-                points.append(pt)
-                count += 1
-
-        vertices, faces = convex_hull_numpy(points)
-
-        i_index = {i: index for index, i in enumerate(vertices)}
-
-        vertices = [points[index] for index in vertices]
-        faces = [[i_index[i] for i in face] for face in faces]
-        faces = unify_cycles(vertices, faces)
-
-        mesh = Mesh.from_vertices_and_faces(vertices, faces)
-
-        viewer = MeshViewer(mesh)
-
-        viewer.setup()
-        viewer.show()
+    >>>
 
     """
     points = asarray(points)
@@ -141,42 +101,3 @@ if __name__ == "__main__":
 
     import doctest
     doctest.testmod(globs=globals())
-
-    # # todo: distinguish between vertices of hull and internal vertices
-
-    # import random
-
-    # from compas.geometry import distance_point_point
-
-    # from compas.datastructures import Mesh
-    # from compas.datastructures import mesh_unify_cycles
-    # from compas_viewers import MeshViewer
-
-    # radius = 5
-    # origin = (0., 0., 0.)
-    # count = 0
-    # points = []
-
-    # while count < 1000:
-    #     x = (random.random() - 0.5) * radius * 2
-    #     y = (random.random() - 0.5) * radius * 2
-    #     z = (random.random() - 0.5) * radius * 2
-    #     pt = x, y, z
-
-    #     if distance_point_point(origin, pt) <= radius:
-    #         points.append(pt)
-    #         count += 1
-
-    # vertices, faces = convex_hull_numpy(points)
-
-    # i_index = {i: index for index, i in enumerate(vertices)}
-
-    # vertices = [points[index] for index in vertices]
-    # faces = [[i_index[i] for i in face] for face in faces]
-
-    # mesh = Mesh.from_vertices_and_faces(vertices, faces)
-    # mesh_unify_cycles(mesh)
-
-    # viewer = MeshViewer()
-    # viewer.mesh = mesh
-    # viewer.show()